feat: add file deletion

+++ api/src/services/files.py
@@ -6,6 +6,7 @@ from typing import List
 
 from fastapi import HTTPException
 from sqlmodel import Session
+from botocore.exceptions import BotoCoreError, ClientError
 
 from src.config import settings
 from src.connectors import s3
@@ -82,12 +83,23 @@ def update_annotations(db: Session, file_id: int, data: UpdateFile):
     return db_file
 
 
-def delete_file(db: Session, id: int):
-    db_file = db.query(Files).filter(Files.id == id).first()
+def delete_file(db: Session, file_id: str):
+    db_file = db.query(Files).filter(Files.id == file_id).first()
+    
+    if not db_file:
+        raise ValueError("File not found")
+    
+    filename = f"{db_file.hash}.{db_file.extension}"
+    print(filename)
+    
+    try:
+        s3.delete_file_obj(filename)
+    except (BotoCoreError, ClientError) as e:
+        print(f"Error deleting file from S3: {e}")
+        return None
     db.delete(db_file)
     db.commit()
     return db_file
-
 
 def deleteAllFilesDeployment(db: Session, id: int):
     db_files = db.query(Files).filter(Files.deployment_id == id).all()
